Completed html validation for all html files
createGroup and base html files have some fixes to improve validity.

But user_creation and user_login have the issue of action cannot be empty in their form tags: <form action="" method="post"
Was wondering if we could put some placeholder in there without messing up the website's functionality or if we will be putting something in there in the future, e.g. maybe the action could be updating the database when we do the linking of database to the webpages?