Updated descriptions of our website/group project
The description is very detailed and thorough, originally created for about.html for users to see, but also works much better for readme.md than what was in there before. This is what I understand our project is doing or at least we would want it to do, so if someone can go through it and make sure it's conceptually sound.